requirements.txt
flask-cors==3.0.10
loguru==0.5.3
pip-autoremove==0.9.1
pip-chill==1.0.1
pytest==6.2.4
python-dotenv==0.18.0
uptime==3.0.1
